[configurator] Fix PS2+ communication library to work with regular controller communication
This change allows the configurator app to successfully pause the
controller interface and issue custom commands. This is required since
the custom PS2+ commands (configuration, updates, etc.) operate on the
same communication channel as the standard DualShock controller
interface (e.g. joysticks, buttons). The controller interface has to be
paused before the custom commands and resumed afterwards in order to
avoid interfering.

Previously, I was doing manual SIO2 communication in the same way that
the MX4SIO project was doing (in fact, copied from there). Didn't fully
understand the reason why, but doing it this way broke the controller
interface causing the configurator to stop being able to read the
controller data.

With this change, I've moved to using the standard (X)SIO2MAN library to
do the PS2+ custom commands in the same way that is done by the
official/free controller interface libraries. This got the commands to
successfully work and not break the controller!