[dev] Re-introduce some of the previous unit tests
Also fixed a bug in `firmware/src/controller/state.c` -- passing in `&state` (which is a `controller_state **) instead of just `state` to `controller_state_reset(controller_state *)`. Don't know what this led to, if anything, on real hardware, but it was causing segmentation faults in the native unit tests.

(see #12)