[dev] Undefine __clang__ in VS Code to fix IntelliSense errors
On platforms that alias GCC to Clang (like the macOS system I'm using currently), VS Code will pick up the `__clang__` define which caused IntelliSense errors in some ImGui code. This `.vscode/pre-include.h` file is only included in VS Code operation and will prevent this; it is *not* included in actual compilation of the project.